clean up coyote-derive and move env_overridable into a module
